feat: + property card as separate component
